requirement.txt
django
django-allauth
djangorestframework
markdown
django-filter

django-admin-material-dashboard
pillow
django-crispy-forms
crispy-bootstrap4

gunicorn
ipython

#django-pdfkit
reportlab
python-environ==0.4.54
nodeenv # venv for react and js staffs  https://varunver.wordpress.com/2019/09/14/setup-react-and-node-js-on-virtualenv/
django-cors-headers


celery
redis
django-celery-results
postmark
django-environ

django-social-share
pdfkit
openpyxl
twilio
drf-yasg
